requirements.txt
# IDP Medical Agent – Intelligent Document Parsing for healthcare facilities
# Python 3.10+

# Orchestration & agents
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
crewai>=0.40.0
crewai-tools>=0.14.0

# Document loading & RAG
llama-index>=0.10.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-readers-file>=0.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0

# API (for frontend / Lovable)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
